delt routes opp i flere filer
delt alle routes opp i filer for bedre oversikt. Fjernet overflødige imports fra api.py og fikset små bugs.
