+++ Requests/GET_R/test_admin.py
@@ -6,15 +6,17 @@ current_directory = os.getcwd()
 import sys
 sys.path.append(os.path.join(current_directory))
 
+#imports requirements
 from Common.Requirements.session_req import require_session
 from Common.Requirements.admin_req import require_admin_account
 
-def test_route(ns):
-    @ns.route('/test')
+#creates test route for admin account
+def test_admin_route(ns):
+    @ns.route('/test_admin')
     class Test(Resource):
-        @ns.doc('test')
+        @ns.doc('test_admin')
         #requirements
         @require_session
-        #@require_admin_account  
+        @require_admin_account  
         def get(self):
             return {"Test": "OK"}
